backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
googlemaps==4.10.0
textblob==0.17.1
python-dotenv==1.0.0
python-multipart==0.0.6
